implemented timing mechanism
added a few lines of code to track how long processing took for multithreading in mandel.c, and multi-processing in mandelseries.c. Added a bool global var (like DBG) to control whether or not that timing info is output.
Could implement an undocumented cmdline param to switch timing on and off