(SEE NOTE) PUT cmd almost fully implemented; several other changes
NOTE: internet was down so had to compile with g++ on the RasPi, so some of the syntax (specifically -> vs. .) might need to be modified for omega)
PUT: just need to finish the time.h stuff, creating the DirectoryEntry and indexing the used blocks should be finished;
DF: done;
added a getIndexOfNextFreeDirEntry() function;
created a rootDirEntries pointer based off the rootDir pointer, which is based off the vfs address;
added some code to initVirtFS() to initialize blank dir entries;
LIST: was seg faulting, fixed that to utilize the new rootDirEntries pointer. Just need to output the entries now, should be quick;
implemented a getInode() function to return a pointer to an inode, which is actually just a pointer to one of the vfs[1-128] elements